added menu, infrastructure, and game variables. fixed shift string
Created a system for storing variables, as well as a way to start the game. Created most of the main menu, just have to bulk it out a bit. Music now only plays over menu, as it gets messed up whenever the program draws to the screen. Also fixed shift string so that it can shift in multiple directions, and added a way to clear out a specific string with a custom character. Split up the timers into a game tick timer and the music timer to make things easier. TODO: we still have to figure out a way to make the graphics draw to the screen better because its scanning is too slow and the refresh is too fast for it currently