Did a lot of stuff
Fixed a lot of small issues, levels are partly implemented, got a way to keep track of score, got a timer, make a neat way of the game ending, as well as a way to replay the game (currently slightly buggy). Lives work correctly, and deaths work 99% as intended. TODO still: Figure out how to get an H to indicate a filled spot in the end, also how to reset the board after each level, and how to populate the board quickly on first starting, since it starts empty. Also we have to get the keypad working. Lastly have to work out some bugs with the board glitching out occasionally on game exit. Finally as a formality I still have to write the game instructions.